splitting access out from mma into new mixin; restructuring dir structure with mixins and decorator file
